macos: remove AppState and unify onto Ghostty.App cross-platform
